+++ src/Instance/InstanceDetails/InstanceDetails.js
@@ -26,7 +26,7 @@ import {
   Row,
   MessageBanner,
   PaneCloseLink,
-  Paneset,
+  Layout,
 } from '@folio/stripes/components';
 import { VersionHistoryButton } from '@folio/stripes-acq-components';
 
@@ -164,6 +164,25 @@ const InstanceDetails = forwardRef(({
     setIsAllExpanded(isExpanded);
   };
 
+  const warningBanners = [
+    {
+      condition: !instance.deleted && instance.staffSuppress && !instance.discoverySuppress,
+      messageId: 'ui-inventory.warning.instance.staffSuppressed',
+    },
+    {
+      condition: !instance.deleted && instance.discoverySuppress && !instance.staffSuppress,
+      messageId: 'ui-inventory.warning.instance.suppressedFromDiscovery',
+    },
+    {
+      condition: !instance.deleted && instance.discoverySuppress && instance.staffSuppress,
+      messageId: 'ui-inventory.warning.instance.suppressedFromDiscoveryAndStaffSuppressed',
+    },
+    {
+      condition: instance.deleted && instance.discoverySuppress && instance.staffSuppress,
+      messageId: 'ui-inventory.warning.instance.setForDeletionAndSuppressedFromDiscoveryAndStaffSuppressed',
+    },
+  ];
+
   return (
     <>
       <Pane
@@ -187,20 +206,18 @@ const InstanceDetails = forwardRef(({
 
         <AccordionStatus ref={ref}>
           <Row>
-            <Col xs={10}>
-              <MessageBanner show={Boolean(instance.staffSuppress && !instance.discoverySuppress)} type="warning">
-                <FormattedMessage id="ui-inventory.warning.instance.staffSuppressed" />
-              </MessageBanner>
-              <MessageBanner show={Boolean(instance.discoverySuppress && !instance.staffSuppress)} type="warning">
-                <FormattedMessage id="ui-inventory.warning.instance.suppressedFromDiscovery" />
-              </MessageBanner>
-              <MessageBanner show={Boolean(instance.discoverySuppress && instance.staffSuppress)} type="warning">
-                <FormattedMessage id="ui-inventory.warning.instance.suppressedFromDiscoveryAndStaffSuppressed" />
-              </MessageBanner>
-            </Col>
-            <Col data-test-expand-all xs={2}>
-              <ExpandAllButton onToggle={onToggle} />
-            </Col>
+            <Layout className="display-flex full flex-align-items-center justify-end">
+              <Col xs={10}>
+                {warningBanners.map(({ condition, messageId }) => (
+                  <MessageBanner marginTop0 key={messageId} show={Boolean(condition)} type="warning">
+                    <FormattedMessage id={messageId} />
+                  </MessageBanner>
+                ))}
+              </Col>
+              <Col data-test-expand-all xs={2}>
+                <ExpandAllButton onToggle={onToggle} />
+              </Col>
+            </Layout>
           </Row>
 
           <InstanceTitle