+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 * Add Version history button and Version history pane to details view of Item. Refs UIIN-3172.
 * Add Version history button and Version history pane to details view of Holdings. Refs UIIN-3171.
 * Add Version history button and Version history pane to details view of Instance. Refs UIIN-3170.
+* Add new ‘Set for deletion’ flag to display on 3rd pane Instance view. Refs UIIN-3191.
 
 ## [12.0.12](https://github.com/folio-org/ui-inventory/tree/v12.0.12) (2025-01-27)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v12.0.11...v12.0.12)